refactory gates package
