+++ engine/src/main/java/it/unibo/arces/wot/sepa/engine/gates/Gate.java
@@ -1,4 +1,4 @@
-package it.arces.wot.sepa.engine.gates;
+package it.unibo.arces.wot.sepa.engine.gates;
 
 import java.util.UUID;
